chore(rename): better use of header prefixes to reflect schema naming
