+++ lib/schemas/index.js
@@ -1,12 +1,12 @@
 'use strict'
 
 module.exports = {
-  cloudflare: require('./cloudflare'),
-  fastly: require('./fastly'),
-  microsoft: require('./microsoft'),
-  nginx: require('./nginx'),
-  rackspace: require('./rackspace'),
-  rfc7239: require('./rfc7239'),
-  xff: require('./xff'),
-  zscaler: require('./zscaler')
+  'cloudflare': require('./cloudflare'),
+  'fastly': require('./fastly'),
+  'microsoft': require('./microsoft'),
+  'rfc7239': require('./rfc7239'),
+  'x-cluster': require('./x-cluster'),
+  'x-forwarded': require('./x-forwarded'),
+  'x-real': require('./x-real'),
+  'z-forwarded': require('./z-forwarded')
 }
